bpf: avoid one csum_l4_replace call and inline csum diff on dport
Extend csum_diff() to handle the case where we have a prior seed, and
compile out the SNAT to loopback entirely since we only ever need it in
the bpf_lxc lb4_xlate() case (netdev lb is compiled with DISABLE_LOOPBACK_LB).

Reorder the dst IP update for csum_l4_replace() to after we rewrote the
ports and inline the csum diff there in order to save one call to the
csum_l4_replace() instead of doing it in l4_modify_port().

Also, small change to reorder the tests on lb{4,6}_svc_is_nodeport() to test
the more commonly used cases first.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>