bpf: Disable relax_verifier()
relax_verifier() is supposed to help the verifier prune branches and
therefore decrease the reported complexity. On 4.19 however, it actually
increases the complexity because it adds instructions in complexity-hot
paths.

After a quick check, it doesn't seem to be required for 4.9 either, so
let's see if we can remove it.

Signed-off-by: Paul Chaignon <paul@cilium.io>